best submission to date
The code dt_sklearn_5features.ipynb, in addition to having one more
feature, is also a big improvement on the workflow. I've introduced
a more formal training / validation separation to allow for
hyperparameter tuning. I've also reworked the submission step to
*hopefully* reduce mistakes.

However ... my computed out of sample scores are not agreeing with
the scores Kaggle computes on the test set. There could be a bug lurking
around.